rename PlaygroundGenerator file
