+++ lib/cocoapods-playgrounds/command.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
-require 'cocoapods-playgrounds/generate'
+require 'cocoapods-playgrounds/playground'
 require 'cocoapods-playgrounds/workspace'
 require 'cocoapods-playgrounds/command/playgrounds'
