+++ lib/cocoapods-playgrounds/playground.rb
